// src/app/models/employee.model.ts
export class Employee {
  [x: string]: any;
  // id?: number = 0;
  firstname: string = '';
  lastname: string = '';
  email: string='';
  birthdate: string = '';
  gender: string = '';
  education: string = '';
  company: string = '';
  department: string='';
  jobExperience: number = 0;
  salary: number = 0;
  profile: string = '';
  }